Junit User Commit

+++ src/main/java/com/onlinepizza/serviceimp/IToppingsServiceImp.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.onlinepizza.entity.Toppings;
+import com.onlinepizza.exception.PizzaToppingsManagementException;
 import com.onlinepizza.repository.ToppingsRepository;
 import com.onlinepizza.service.IToppingsService;
 
@@ -12,6 +13,7 @@ public class IToppingsServiceImp implements IToppingsService{
 
 	@Autowired
 	ToppingsRepository toppingsRepository;
+	
 	@Override
 	public String addToppings(Toppings topping) {
 		
@@ -19,7 +21,14 @@ public class IToppingsServiceImp implements IToppingsService{
 		return "Toppings added";
 	}
 	@Override
-	public String deleteToppingsById(Integer toppingsId) {
+	public String deleteToppingsById(Integer toppingsId) throws PizzaToppingsManagementException
+	{
+		if(toppingsRepository.findById(toppingsId).isEmpty())
+		{
+			throw new PizzaToppingsManagementException ("Toppings not found");
+		}
+		else 
+			
 		toppingsRepository.deleteById(toppingsId);
 		return "Toppings Removed Successfully";
 	}
